Add search functionality to courses and social
Add search bar for social
Add search service to allow sending the search text from the search bar to the courses and the social pages.
Add function to filter the courses.
Add function to filter the forums in social.
Change the way forums are iterated.
Remove deprecated interface for Social Card.